minor changes 1. new widget in homescreen 2. ui update in history page 3. database change in history table 4. datatype change in emi calculator view 5. new constants

+++ lib/logic/logic.dart
@@ -1,7 +1,7 @@
 import 'package:intl/intl.dart';
 
 class LoanCalculator {
-  final int takenAmount;
+  final double takenAmount;
   final double rateOfInterest;
   final String takenDate;
   DateTime? tenureDate;
